removed unused imports, added comments, added session user matching altered user conditionals

+++ routes/list.py
@@ -1,8 +1,9 @@
 import requests
 import random
-from flask import request, render_template, redirect, g, Blueprint, flash
-from models import db, List, Listings, Anime, likes, User
-from forms import EditListForm
+from flask import request, render_template, redirect, g, Blueprint
+from models import db, List, Listings, Anime, likes
+from forms import EditListForm, EditListingForm
+
 
 
 list = Blueprint('list', __name__,template_folder='routes')
@@ -14,6 +15,9 @@ list = Blueprint('list', __name__,template_folder='routes')
 @list.route("/new_list/")
 def new_list():
     """Create an empty new list."""
+
+    if not g.user:
+        return redirect("/") 
 
     list = List(
             title="",
@@ -28,6 +32,7 @@ def new_list():
 
 @list.route("/recommend/", methods=["GET", "POST"])
 def recommend():
+    """Generate a list of recommendations."""
 
     if request.method == "GET":
 
@@ -78,6 +83,7 @@ def recommend():
         db.session.add(list)
         db.session.commit()
 
+        # randomize if returned recommendations is greater than 10 count
         if len(resp.json()['data']) > 10:
             random_array = random.sample(range(len(resp.json()['data'])), 10)
             for num in random_array:
@@ -88,10 +94,12 @@ def recommend():
                 )
                 db.session.add(listing)
                 db.session.commit()
+
         return redirect(f"/list/{list.list_id}")
 
 @list.route("/list/<int:list_id>/", methods=["GET", "POST"])
 def show_list(list_id):
+    """Show list owned by user."""
 
     list = List.query.get_or_404(list_id)
 
@@ -99,19 +107,25 @@ def show_list(list_id):
     # prepopulate the edit_form textarea
     edit_form.list_description.data = list.description
 
-    likes_count = db.session.query(likes).filter_by(list_id=list_id).count()
+    edit_listing_form = EditListingForm()
 
+    # Check if user has liked before 
+    likes_count = db.session.query(likes).filter_by(list_id=list_id).count()
 
     if g.user:
         user_has_liked = db.session.query(likes).filter_by(user_id=g.user.user_id, list_id=list_id).first()
-
     else:
         user_has_liked = False
 
     if request.method == "GET":
-        return render_template("/list/list.html", list=list, edit_form=edit_form, likes_count=likes_count, user_has_liked=user_has_liked)
+        return render_template("/list/list.html", list=list, edit_form=edit_form, edit_listing_form=edit_listing_form, likes_count=likes_count, user_has_liked=user_has_liked)
     
+    # If built in edit list details form is submitted
     if request.method == "POST":
+
+        if g.user.user_id != list.user_id:
+            return redirect(f"/list/{list_id}")
+
         list.title = request.form.get('list_title')
         list.description = request.form.get('list_description')
 
@@ -121,13 +135,19 @@ def show_list(list_id):
 
 @list.route("/list/<int:list_id>/delete/", methods=["GET", "POST"])
 def delete_list(list_id):
+    """Delete list."""
 
     list = List.query.get_or_404(list_id)
 
     if not g.user:
         return redirect(f"/list/{list_id}")
+
+    if g.user.user_id != list.user_id:
+        return redirect(f"/list/{list_id}")
+
     
     if request.method == "POST":  
+
         Listings.query.filter_by(list_id=list_id).delete()
         db.session.delete(list)
         db.session.commit()
@@ -135,18 +155,20 @@ def delete_list(list_id):
         
 @list.route("/list/<int:list_id>/search/", methods=["GET","POST"])
 def search(list_id):
+    """Search function."""
 
     list = List.query.get_or_404(list_id)
 
     if not g.user:
         return redirect(f"/list/{list_id}") 
     
-
+    # preserve search_input through page swaps
     search_input = request.args.get('search_input')
     if search_input == None:
         # Setting default to search_input (seems to fix first search returning None)
-        search_input = "Bocchi"
+        search_input = "Bocchi The Rock"
 
+    # retrieve current page input
     try:
         page = int(request.args.get('page'))
     except:
@@ -167,20 +189,24 @@ def search(list_id):
 
 @list.route("/list/<int:list_id>/add", methods=["GET", "POST"])
 def add_to_list(list_id):
+    """Add an anime listing to our list."""
 
     list = List.query.get_or_404(list_id)
 
-    mal_id = request.args.get('mal_id')
-
-    print(list, mal_id)
+    if g.user.user_id != list.user_id:
+        return redirect(f"/list/{list_id}")
 
     if not g.user:
         return redirect(f"/list/{list_id}") 
     
+    mal_id = request.form.get('mal_id')
+
+    # Checks if anime is already in our database (lowers how many API requests we have to make)
     if Anime.query.get(mal_id) == None:
 
         resp = requests.get(f"https://api.jikan.moe/v4/anime/{mal_id}")
 
+        # Save anime to our database for future use
         anime = Anime(
             anime_id=resp.json()['data']['mal_id'],
             anime_image_url=resp.json()['data']['images']['jpg']['image_url'],
@@ -203,7 +229,6 @@ def add_to_list(list_id):
     else: 
 
         anime = Anime.query.get(mal_id)
-        print(anime)
         listing = Listings(
             list_id=list_id,
             anime_id = anime.anime_id,
@@ -216,17 +241,21 @@ def add_to_list(list_id):
     return redirect(f"/list/{list_id}")
 
 
-@list.route("/listing/<int:listing_id>/edit/", methods=["GET","POST"])
+@list.route("/edit-listing/<int:listing_id>", methods=["GET","POST"])
 def edit_listing(listing_id):
+    """Edit a listing."""
 
     listing = Listings.query.get_or_404(listing_id)
 
     if not g.user:
         return redirect(f"/list/{listing.lists.list_id}")
     
+    if g.user.user_id != listing.lists.user_id:
+        return redirect("/")
+    
     if request.method == "POST":
 
-        listing_description = request.form.get('listing-description')
+        listing_description = request.form.get('listing_description')
         listing.listing_description = listing_description
         db.session.add(listing)
         db.session.commit()
@@ -234,18 +263,26 @@ def edit_listing(listing_id):
     
 @list.route("/listing/<int:listing_id>/delete/", methods=["GET","POST"])
 def delete_listing(listing_id):
+    """Delete a listing."""
 
     listing = Listings.query.get_or_404(listing_id)
+    # Preserve list_id in variable
+    list_id = listing.lists.list_id
 
     if not g.user:
-        return redirect(f"/list/{listing.lists.list_id}")   
+        return redirect(f"/list/{listing.lists.list_id}")  
     
-    db.session.delete(listing)
-    db.session.commit()
-    return redirect(f"/list/{listing.lists.list_id}")
+    if g.user.user_id != list_id:
+        return redirect(f"/list/{list_id}")
+     
+    if request.method == "POST":
+        db.session.delete(listing)
+        db.session.commit()
+        return redirect(f"/list/{list_id}")
 
 @list.route("/list/<int:list_id>/like", methods=["GET","POST"])
 def like_list(list_id):
+    """Like a list."""
 
     list = List.query.get_or_404(list_id)
